Move all tests out of test_ambassador.py into t_ files, to make it easier to pick and choose in development.
