+++ ambassador/tests/t_mappingtests.py
@@ -1,0 +1,228 @@
+from abstract_tests import MappingTest, AmbassadorTest, ServiceType, OptionTest, HTTP
+from harness import variants, Query
+from test_ambassador import unique
+
+
+class SimpleMapping(MappingTest):
+
+    parent: AmbassadorTest
+    target: ServiceType
+
+    @classmethod
+    def variants(cls):
+        for st in variants(ServiceType):
+            yield cls(st, name="{self.target.name}")
+
+            for mot in variants(OptionTest):
+                yield cls(st, (mot,), name="{self.target.name}-{self.options[0].name}")
+
+            yield cls(st, unique(v for v in variants(OptionTest)
+                                 if not getattr(v, "isolated", False)), name="{self.target.name}-all")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v1
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: http://{self.target.path.fqdn}
+""")
+
+    def queries(self):
+        yield Query(self.parent.url(self.name + "/"))
+        yield Query(self.parent.url(f'need-normalization/../{self.name}/'))
+
+    def check(self):
+        for r in self.results:
+            if r.backend:
+                assert r.backend.name == self.target.path.k8s, (r.backend.name, self.target.path.k8s)
+                assert r.backend.request.headers['x-envoy-original-path'][0] == f'/{self.name}/'
+
+
+class HostHeaderMapping(MappingTest):
+
+    parent: AmbassadorTest
+
+    @classmethod
+    def variants(cls):
+        for st in variants(ServiceType):
+            yield cls(st, name="{self.target.name}")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v1
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: http://{self.target.path.fqdn}
+host: inspector.external
+""")
+
+    def queries(self):
+        yield Query(self.parent.url(self.name + "/"), expected=404)
+        yield Query(self.parent.url(self.name + "/"), headers={"Host": "inspector.internal"}, expected=404)
+        yield Query(self.parent.url(self.name + "/"), headers={"Host": "inspector.external"})
+
+
+class WebSocketMapping(MappingTest):
+
+    parent: AmbassadorTest
+
+    @classmethod
+    def variants(cls):
+        for st in variants(ServiceType):
+            yield cls(st, name="{self.target.name}")
+
+    def config(self):
+        yield self, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: echo.websocket.org:80
+host_rewrite: echo.websocket.org
+use_websocket: true
+""")
+
+    def queries(self):
+        yield Query(self.parent.url(self.name + "/"), expected=404)
+
+        yield Query(self.parent.url(self.name + "/"), expected=101, headers={
+            "Connection": "Upgrade",
+            "Upgrade": "websocket",
+            "sec-websocket-key": "DcndnpZl13bMQDh7HOcz0g==",
+            "sec-websocket-version": "13"
+        })
+
+        yield Query(self.parent.url(self.name + "/", scheme="ws"), messages=["one", "two", "three"])
+
+    def check(self):
+        assert self.results[-1].messages == ["one", "two", "three"]
+
+
+class TLSOrigination(MappingTest):
+
+    parent: AmbassadorTest
+    definition: str
+
+    IMPLICIT = """
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: https://{self.target.path.fqdn}
+"""
+
+    EXPLICIT = """
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: {self.target.path.fqdn}
+tls: true
+"""
+
+    @classmethod
+    def variants(cls):
+        for v in variants(ServiceType):
+            for name, dfn in ("IMPLICIT", cls.IMPLICIT), ("EXPLICIT", cls.EXPLICIT):
+                yield cls(v, dfn, name="{self.target.name}-%s" % name)
+
+    def init(self, target, definition):
+        MappingTest.init(self, target)
+        self.definition = definition
+
+    def config(self):
+        yield self.target, self.format(self.definition)
+
+    def queries(self):
+        yield Query(self.parent.url(self.name + "/"))
+
+    def check(self):
+        for r in self.results:
+            assert r.backend.request.tls.enabled
+
+
+class HostRedirectMapping(MappingTest):
+    parent: AmbassadorTest
+    target: ServiceType
+
+    def init(self):
+        MappingTest.init(self, HTTP())
+
+    def config(self):
+        yield self.target, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: foobar.com
+host_redirect: true
+""")
+
+    def queries(self):
+        yield Query(self.parent.url(self.name + "/anything?itworked=true"), expected=301)
+
+    def check(self):
+        assert self.results[0].headers['Location'] == [
+            self.format("http://foobar.com/{self.name}/anything?itworked=true")
+        ]
+
+
+class CanaryMapping(MappingTest):
+
+    parent: AmbassadorTest
+    target: ServiceType
+    canary: ServiceType
+    weight: int
+
+    @classmethod
+    def variants(cls):
+        for v in variants(ServiceType):
+            for w in (10, 50):
+                yield cls(v, v.clone("canary"), w, name="{self.target.name}-{self.weight}")
+
+    def init(self, target: ServiceType, canary: ServiceType, weight):
+        MappingTest.init(self, target)
+        self.canary = canary
+        self.weight = weight
+
+    def config(self):
+        yield self.target, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}
+prefix: /{self.name}/
+service: http://{self.target.path.fqdn}
+""")
+        yield self.canary, self.format("""
+---
+apiVersion: ambassador/v0
+kind:  Mapping
+name:  {self.name}-canary
+prefix: /{self.name}/
+service: http://{self.canary.path.fqdn}
+weight: {self.weight}
+""")
+
+    def queries(self):
+        for i in range(100):
+            yield Query(self.parent.url(self.name + "/"))
+
+    def check(self):
+        hist = {}
+
+        for r in self.results:
+            hist[r.backend.name] = hist.get(r.backend.name, 0) + 1
+
+        canary = 100*hist.get(self.canary.path.k8s, 0)/len(self.results)
+        # main = 100*hist.get(self.target.path.k8s, 0)/len(self.results)
+
+        assert abs(self.weight - canary) < 25, (self.weight, canary)
